divide code to modules and import git only when required
